Refactor stock query imports: Move `queryStock.js` to `lib/stock` for improved structure and update all imports accordingly. Add minor adjustment to global styles.
