+++ app/globals.css
@@ -74,6 +74,7 @@
 .dashboard-table-heading {
     @apply px-6 py-3 text-left text-xs font-medium text-gray-300 uppercase tracking-wider;
 }
+
 @keyframes indeterminate {
     0% {
         transform: translateX(-100%);
