+++ lib/queryPurchaseOrders.js
@@ -3,7 +3,7 @@
 import {createClient} from "@/lib/supabase/server";
 import {getCurrentOrganizationId} from "@/lib/queryOrganizations";
 import {getCurrentLoggedInUser} from "@/lib/queryUsers";
-import {adjustStock} from "@/lib/queryStock";
+import {adjustStock} from "@/lib/stock/queryStock";
 
 export async function getPurchaseOrdersForCurrentOrganization(options = {}) {
     const supabase = await createClient();
