+++ lib/stock/queryStock.js
